move UportDIDDocument class to its own file and add comments to the now deprecated profile model
